restructure components with functional composition
