+++ src/components/button/__test__/index.test.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import { shallow } from 'enzyme'
+import Button from '../'
+
+describe('<Button /> component index', () => {
+  it('expects component to match snapshot', () => {
+    const wrapper = shallow(<Button />).dive()
+    expect(wrapper).toMatchSnapshot()
+  })
+})
