+++ src/components/claim/__test__/View.test.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import Claim from './Claim'
+import View from '../View'
 
-describe('<Claim />', () => {
+describe('<Claim /> component View', () => {
   it('renders without crashing', () => {
-    const wrapper = shallow(<Claim />)
+    const wrapper = shallow(<View />)
     expect(wrapper.find('.Claim')).toHaveLength(1)
   })
 })
