+++ src/components/claim/__test__/index.test.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import { shallow } from 'enzyme'
+import Claim from '../'
+
+describe('<Claim /> component index', () => {
+  it('expects component to match snapshot', () => {
+    const wrapper = shallow(<Claim />).dive()
+    expect(wrapper).toMatchSnapshot()
+  })
+})
