+++ src/components/claim/styles.js
