+++ src/components/footer/Logic.js
@@ -1,0 +1,3 @@
+import { withCallRoute } from '../higher-order/with-call-route'
+
+export default withCallRoute
