+++ src/components/footer/__test__/View.test.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import Footer from './Footer'
+import Footer from '../View'
 
-describe('<Button />', () => {
+describe('<Footer /> component View', () => {
   it('renders without crashing', () => {
     const wrapper = shallow(<Footer />)
     expect(wrapper.find('.Footer')).toHaveLength(1)
