+++ src/components/footer/index.js
@@ -1,0 +1,5 @@
+import { compose } from 'recompose'
+import View from './View'
+import Logic from './Logic'
+
+export default compose(Logic)(View)
