+++ src/components/footer/styles.js
