+++ src/components/header/__test__/index.test.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import { shallow } from 'enzyme'
+import Header from '../'
+
+describe('<Header /> component index', () => {
+  it('renders without crashing', () => {
+    const wrapper = shallow(<Header />)
+    expect(wrapper).toMatchSnapshot()
+  })
+})
