+++ src/components/higher-order/with-call-route.js
@@ -1,0 +1,7 @@
+import { compose, withProps } from 'recompose'
+
+export const withCallRoute = withProps(({ route }) => ({
+  callRoute: () => (window.location.href = route),
+}))
+
+export default compose(withCallRoute)
