+++ src/components/navbar/__test__/View.test.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import Navbar from './Navbar'
+import Navbar from '../View'
 
-describe('<Navbar />', () => {
+describe('<Navbar /> component View', () => {
   it('renders without crashing', () => {
     const wrapper = shallow(<Navbar />)
     expect(wrapper.find('.Navbar')).toHaveLength(1)
