+++ src/components/navbar/index.js
@@ -1,0 +1,3 @@
+import View from './View'
+
+export default View
