+++ src/components/skill-card/__test__/View.test.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import SkillCard from './SkillCard'
+import SkillCard from '../View'
 
-describe('<SkillCard />', () => {
+describe('<SkillCard /> component View', () => {
   it('renders without crashing', () => {
     const wrapper = shallow(<SkillCard />)
     expect(wrapper.find('.SkillCard')).toHaveLength(1)
