+++ src/components/skill-card/styles.js
